Refactor the MediaGallery plugin responsible for operations after image or diretory deleted
